requirements.txt
# 核心依赖
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.14.0

# 训练相关
accelerate>=0.24.0
peft>=0.6.0  # 用于LoRA等参数高效微调方法
deepspeed>=0.11.0  # 分布式训练（可选）

# 数据处理
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0

# 配置管理
pyyaml>=6.0
omegaconf>=2.3.0

# 实验追踪
wandb>=0.15.0  # 实验追踪（可选）
tensorboard>=2.14.0

# 评估
matplotlib>=3.7.0
seaborn>=0.12.0

# 推理（可选）
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# 开发工具
jupyter>=1.0.0
ipykernel>=6.25.0

# 代码质量
pytest>=7.4.0
black>=23.7.0
ruff>=0.0.280
mypy>=1.5.0